Nmrc

Documentation for Nmrc.

Nmrc.PML

Nmrc.allocate_matrices

Nmrc.apply_all_bds!

Nmrc.assemble_A0

Nmrc.assemble_A1

Nmrc.assemble_A2

Nmrc.compute_scaling

Nmrc.coord_transform

Nmrc.get_width

Nmrc.setup_bdcs

Nmrc.setup_dofs

Nmrc.setup_fevs

Nmrc.setup_grid

Nmrc.PML — Type

Pml

Information on PML layer and coordinate transformation.

Nmrc.allocate_matrices — Method

allocate_matrices(dof::DofHandler, cst::ConstraintHandler)

Allocate the complex-valued stiffness matrices.

Nmrc.apply_all_bds! — Method

apply_all_bds!(K::Union{SparseMatrixCSC, Symmetric}, ch::ConstraintHandler)
apply_all_bds!(KK::Union{SparseMatrixCSC, Symmetric}, f::AbstractVector, ch::ConstraintHandler, applyzero::Bool = false)

Impose the periodic boundary condition and the Dirichlet boundary condition on the matrices.

Nmrc.assemble_A0 — Method

assemble_A0(cv, dh, A₀, medium, p, k)

Assemble the zero order term in the quadratic eigenvalue problem.

Nmrc.assemble_A1 — Method

assemble_A1(cv::CellValues, dh::DofHandler, A₁::SparseMatrixCSC, p::PML)

Assemble the first order term in the quadratic eigenvalue problem.

Nmrc.assemble_A2 — Method

assemble_A2(cv::CellValues, dh::DofHandler, A₂::SparseMatrixCSC, p::PML)

Assemble the second order term in the quadratic eigenvalue problem.

Nmrc.compute_scaling — Method

compute_scaling(A₀::M, A₁::M, A₂::M; scaling=:nothing::Symbol) where {T,M<:AbstractMatrix{T}}

TBW

Nmrc.coord_transform — Method

coord_transform(x₂, p::PML)

Coordinate transformation of PML method.

Nmrc.get_width — Method

get_width(p::PML)

Get the width of the PML layer.

Nmrc.setup_bdcs — Method

setup_bdcs(dof::DofHandler, d)

Set the periodic boundary condition ("left" and "right") and Dirichelt boundary condition ("bottom" and "top").

Nmrc.setup_dofs — Method

setup_dofs(grid::Grid, ip)

Setup degree of freedoms.

Nmrc.setup_fevs — Method

setup_fevs(ip)

Define quadrature rules on Reference triangle and setup FE values.

Nmrc.setup_grid — Method

setup_grid(;d=2π, ĥ=1.5, δ=2.0, lc=0.5, lp=0.5, vtk=false)

Generate a mesh by julia interface for Gmsh and read the .msh file by FerriteGmsh.

Arguments

d: the period of the periodic layer

ĥ: the start of the PML layer

δ: the height of the PML layer

lc: mesh size near the periodic layer

lp: mesh size in the PML layer

vtk: write .vtk file, default is false

Points

 5 ------------- 4
 |               |
 6               3
 |               |
 1 ------------- 2

Lines

 . ----- l4 ----- . 
 l5               l3
 .                .
 l6               l2 
 . ----- l1 ----- .

Misc

We can generate a *.vtk file by replace the do end block with gmsh.write(*.vtk)