Nmrc

Documentation for Nmrc.

Nmrc.PML

Nmrc.allocate_matrices

Nmrc.assemble_A0

Nmrc.assemble_A1

Nmrc.assemble_A2

Nmrc.compute_scaling

Nmrc.coord_transform

Nmrc.get_width

Nmrc.setup_bdcs

Nmrc.setup_dofs

Nmrc.setup_fevs

Nmrc.solve_qm

Nmrc.PML — Type

Pml

Information on PML layer and coordinate transformation.

Nmrc.allocate_matrices — Method

allocate_matrices(dof::DofHandler, cst::ConstraintHandler)

Allocate the complex-valued stiffness matrices.

Nmrc.assemble_A0 — Method

assemble_A0(cv, dh, A₀, medium, p, k)

The zero order term in the quadratic eigenvalue problem

Nmrc.assemble_A1 — Method

assemble_A1(cv, dh, A₁, p)

The first order term in the quadratic eigenvalue problem

Nmrc.assemble_A2 — Method

assemble_A2(cv, dh, A₂, p)

The second order term in the quadratic eigenvalue problem

Nmrc.compute_scaling — Method

compute_scaling(A₀::M, A₁::M, A₂::M; scaling=:nothing::Symbol) where {T,M<:AbstractMatrix{T}}

TBW

Nmrc.coord_transform — Method

coord_transform(x₂, p::PML)

Coordinate transformation of PML method.

Nmrc.get_width — Method

get_width(p::PML)

Get the width of the PML layer.

Nmrc.setup_bdcs — Method

setup_bdcs(dof::DofHandler, d)

Set the periodic boundary condition ("left" and "right") and Dirichelt boundary condition ("bottom" and "top").

Nmrc.setup_dofs — Method

setup_dofs(grid::Grid, ip)

Setup degree of freedoms.

Nmrc.setup_fevs — Method

setup_fevs(ip)

Define quadrature rules on Reference triangle and setup FE values.

Nmrc.solve_qm — Method

solve_qm(cv::CellValues, dof::DofHandler, cst::ConstraintHandler, A₀::SparseMatrixCSC, A₁::SparseMatrixCSC, A₂::SparseMatrixCSC)

Get the exceptional values (quasi-momentum α) by solving the eigenvalue problems generated by FEM.