Nmrc

Documentation for Nmrc.

Usage

Nmrc.jl can be installed by opening a Julia REPL and typing:

]add https://github.com/peakfind/Nmrc.jl.git

Methods Implemented in Nmrc.jl

Use PML and FEM to estimate the exceptional values (quasi momentum).

FEM with transparent boundary condition (DtN map).

TBD

High order quadrature rules on facets may change the sparse pattern of the vector $\Theta^n$.

Nmrc.Incident

Nmrc.Incident

Nmrc.PML

Nmrc.allocate_matrices

Nmrc.allocate_stiff_matrix

Nmrc.apply_all_bds!

Nmrc.assemble_A

Nmrc.assemble_A0

Nmrc.assemble_A1

Nmrc.assemble_A2

Nmrc.assemble_load

Nmrc.assemble_tbc

Nmrc.beta_n

Nmrc.compute_coef!

Nmrc.compute_scaling

Nmrc.coord_transform

Nmrc.dofs_on_dtn

Nmrc.get_width

Nmrc.periodic_cell

Nmrc.setup_bcs

Nmrc.setup_bdcs

Nmrc.setup_dofs

Nmrc.setup_fevs

Nmrc.setup_grid

Nmrc.setup_vals

Nmrc.sub_preserve_structure

Nmrc.Incident — Type

Incident

Some parameters related to the incident field.

Parameters

k: the wave number

θ: the incident field

α: α = k * sin(θ)

β: β = k * cos(θ)

Nmrc.Incident — Method

Incident(k::Float64, θ::Float64)

Define the incident field with the wavenumber k and the incident field θ.

Nmrc.PML — Type

Pml

Information on PML layer and coordinate transformation.

Nmrc.allocate_matrices — Method

allocate_matrices(dof::DofHandler, cst::ConstraintHandler)

Allocate the complex-valued stiffness matrices.

Nmrc.allocate_stiff_matrix — Method

allocate_stiff_matrix(dofhandler::DofHandler, csthandler::ConstraintHandler, dofs)

Create a sparse pattern for the stiffness matrix. We need to add extra entries due to the DtN map by hand.

Nmrc.apply_all_bds! — Method

apply_all_bds!(K::Union{SparseMatrixCSC, Symmetric}, ch::ConstraintHandler)
apply_all_bds!(KK::Union{SparseMatrixCSC, Symmetric}, f::AbstractVector, ch::ConstraintHandler, applyzero::Bool = false)

Impose the periodic boundary condition and the Dirichlet boundary condition on the matrices.

Nmrc.assemble_A — Method

assemble_A(cv::CellValues, dh::DofHandler, A::SparseMatrixCSC, inc::Incident)

Nmrc.assemble_A0 — Method

assemble_A0(cv, dh, A₀, medium, p, k)

Assemble the zero order term in the quadratic eigenvalue problem.

Nmrc.assemble_A1 — Method

assemble_A1(cv::CellValues, dh::DofHandler, A₁::SparseMatrixCSC, p::PML)

Assemble the first order term in the quadratic eigenvalue problem.

Nmrc.assemble_A2 — Method

assemble_A2(cv::CellValues, dh::DofHandler, A₂::SparseMatrixCSC, p::PML)

Assemble the second order term in the quadratic eigenvalue problem.

Nmrc.assemble_load — Method

assemble_load(fv::FacetValues, dh::DofHandler, facetset, f, inc::Incident, height)

Assemble the load vector due to the incident field.

Nmrc.assemble_tbc — Method

assemble_tbc(fv::FacetValues, dh::DofHandler, inc::Incident, F::SparseMatrixCSC, facetset, N, dofsDtN)

Assemble the TBC matrix.

Nmrc.beta_n — Method

beta_n(inc::Incident, n)

Compute the βₙ.

Nmrc.compute_coef! — Method

compute_coef!(fv::FacetValues, dh::DofHandler, θ::SparseVector, facetset, n)

Compute Θⁿ on the facetset. Actually the computation of the TBC matrix reduces to the computation of the vector Θⁿ.

Nmrc.compute_scaling — Method

compute_scaling(A₀::M, A₁::M, A₂::M; scaling=:nothing::Symbol) where {T,M<:AbstractMatrix{T}}

TBW

Nmrc.coord_transform — Method

coord_transform(x₂, p::PML)

Coordinate transformation of PML method.

Nmrc.dofs_on_dtn — Method

dofs_on_dtn(dh::DofHandler, field::Symbol, facetset)

Extract the global indices of Dofs associated to the artificial boundary.

Nmrc.get_width — Method

get_width(p::PML)

Get the width of the PML layer.

Nmrc.periodic_cell — Function

periodic_cell(lc=0.5; height)

Generate a mesh for the periodic cell with period 2π and height height.

Arguments

lc: the mesh size near points

height: the height of the periodic cell

Points

 4 ------------ 3
 |              |
 |              |
 1 ------------ 2

Lines

 . ---- l3 ---- .
 l4             l2
 . ---- l1 ---- .

Nmrc.setup_bcs — Method

setup_bcs(dh::DofHandler)

Set the periodic boundary condition ("left" and "right") and Dirichelt boundary condition ("bottom").

Nmrc.setup_bdcs — Method

setup_bdcs(dof::DofHandler, d)

Set the periodic boundary condition ("left" and "right") and Dirichelt boundary condition ("bottom" and "top").

Nmrc.setup_dofs — Method

setup_dofs(grid::Grid, ip)

Setup degree of freedoms.

Nmrc.setup_fevs — Method

setup_fevs(ip)

Define quadrature rules on Reference triangle and setup FE values.

Nmrc.setup_grid — Method

setup_grid(;d=2π, ĥ=1.5, δ=2.0, lc=0.5, lp=0.5, vtk=false)

Generate a mesh by julia interface for Gmsh and read the .msh file by FerriteGmsh.

Arguments

d: the period of the periodic layer

ĥ: the start of the PML layer

δ: the height of the PML layer

lc: mesh size near the periodic layer

lp: mesh size in the PML layer

vtk: write .vtk file, default is false

Points

 5 ------------- 4
 |               |
 6               3
 |               |
 1 ------------- 2

Lines

 . ----- l4 ----- . 
 l5               l3
 .                .
 l6               l2 
 . ----- l1 ----- .

Misc

We can generate a *.vtk file by replace the do end block with gmsh.write(*.vtk)

Nmrc.setup_vals — Method

setup_vals(ip)

Set up CellValues and FacetValues by using interpolation ip. Here we need to define FacetValues because the load vector and the TBC matrix contains integral on the boundary.

Nmrc.sub_preserve_structure — Method

sub_preserve_structure(A::SparseMatrixCSC, B::SparseMatrixCSC)

Subtract two sparse matrix without changing the sparse pattern. A and B should have the same sparse pattern.